requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-dotenv
psycopg2-binary
python-multipart